requirements.txt
Django==2.1.4
django-cors-headers==3.2.0
djangorestframework==3.9.0
html5lib==1.0.1
Pillow==6.2.1
psycopg2==2.8.3
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.32
six==1.13.0
webencodings==0.5.1
xhtml2pdf==0.2.3
